Unconstrained UK formulation.
Working version of the unconstrained UK formulation. I've defined a few classes, the code should probably be reorganized. It seems to yield the proper q(t) but it should be tested with the modeshapes to be sure. I have only considered a simple ramp excitation for now.
TODO: Obtain y(x,t) by multiplying by the modeshapes. Add the constraints.